+++ java/FirstConsole/src/products/ProductAbstract.java
@@ -2,10 +2,10 @@ package products;
 
 public abstract class ProductAbstract {
     private final String name;
-    private final int makePrice;
-    private final int sellPrice;
+    private final double makePrice;
+    private final double sellPrice;
 
-    public ProductAbstract(String name, int makePrice, int sellPrice) {
+    public ProductAbstract(String name, double makePrice, double sellPrice) {
         this.name = name;
         this.makePrice = makePrice;
         this.sellPrice = sellPrice;
@@ -15,11 +15,11 @@ public abstract class ProductAbstract {
         return this.name;
     }
 
-    public int getSellPrice(){
+    public double getSellPrice(){
         return sellPrice;
     }
 
-    public int getMakePrice(){
+    public double getMakePrice(){
         return makePrice;
     }
 }
